Abort a failed merge, to allow testing to continue.
Previously, if the merge failed, this merge script would fail, and the
whole build would fail without doing any useful testing. By aborting
the merge, testing can at least continue on the unmerged changes,
testing the branch in isolation.

NB. later, after the merge conflicts are resolved, testing will run
again and do the testing of the merge commit, so this doesn't weaken
guarantees of the testing of the final product.